// cosmowcl/src/app/shared/models/api/cosmobusiness/esecuzioneMultiplaApprovaRequest.d.ts
/*
 * Copyright CSI-Piemonte - 2023
 * SPDX-License-Identifier: GPL-3.0-or-later
 */
/**
 * cosmobusiness
 * Api per cosmobusiness
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AttivitaEseguibileMassivamente } from './attivitaEseguibileMassivamente.d';


/**
 * Model con le informazioni per EsecuzioneMultiplaApprovaRequest
 */
export interface EsecuzioneMultiplaApprovaRequest { 
    note?: string;
    approvazione: boolean;
    tasks: Array<AttivitaEseguibileMassivamente>;
    mandareAvantiProcesso: boolean;
}